JSUP-12627: Not working counter of Writer's Block
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@22300 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ cgi-bin/LJ/Entry.pm
@@ -2382,7 +2382,10 @@ sub set_logprop {
     $u->do("DELETE FROM logprop2 WHERE journalid=? AND jitemid=? ".
            "AND propid IN ($del_ids)", undef, $u->{'userid'}, $jitemid) if $del_ids;
 
-    LJ::MemCache::delete([$uid,"logprop2:$uid:$jitemid"]) if $kill_mem;
+    if ($kill_mem) {
+        LJ::MemCache::delete([$uid,"logprop2:$uid:$jitemid"]) if $kill_mem;
+        $singletons{$uid}->{$jitemid} = undef if $singletons{$uid};
+    }
 }
 
 # <LJFUNC>
